Major optimisation changes
Made substitute functions to Numpy functions and removed a few stylistic details for better performance in the Radar Chart. Explicitly deleted some variables and collected garbage on the end of the Radar Chart, as the interpreter would not clean them up automatically and they'd end up stacking up RAM usage.